lidar_script/requirements.txt
pyserial
laspy
numpy